Increased button functionality
made some adjustments to the button. It is now slightly bigger and the trigger box has been adjusted to only trigger when the player is standing on it so the player can jump to press the button.

The button now has some extra functionality that can be accessed from the details Tab
- Button can now be set to toggle on and off when pressed instead of only on when pressed
-Button can now be set to require a prerequisite to be functional (said prerequisite will need to cast to the button to enable functionality)
-Button can now be set to have a set number of uses before losing functionality.