Buttons Revamped
Added BaseButton blueprint that handles interactions between Activators
Added ActivatorBase to use for handling communication with buttons (Anything that needs to be interacted by buttons needs to inherit from this blueprint).
Created a new pressure switch that functions the same as the previous but now can account for multiple players on the switch.

added interaction switch that is activated by pressing "F" on key board or X on a xbox controler.