Climbing Implementation Beginning
Realised BP_WallClimb wont work with side scroll so duplicated it and made BP_WallClimb_SideScroll. Need to re-make it to work by casting the new character and also limit side movements so it cant climb sideways like the 3D one

Flipped side scroll map around so that characters face camera (was done on a demo before but not on this branch yet)

Looked at stopping camera interact with objects but I don't think it does anyway, did some testing and it seems fine.