Added Other Colours for Object Highlighting
The colour of the highlight can now be selected, this is so its possible to differentiate if an object can be repaired or destroyed based on the colour.

It can be changed with node "Set Custom Depth Stencil Value"
The values range from 0 - 2, and the colours of those values can be changed in "MI_ObjectOutline". Currently 0 = blue, 1 = range and 2 = red.
Box, Box2 and Box3 have these new highlight colours applied.

BP_Box has got an added cube collision called "HighlightRadius" which surrounds the area filled by floating debris so that the repair pot has a large area to look at.
It currently doesn't function in BlockOutTestMap because I don't want to create conflicts there.
In order to make it work, in scene only:
- Select all the fragments and change the component tag to "nameR" where name can be anything
- Select HighlightRadius and change the component tag to "nameB" where name can be anything
HighlightRadius is enabled if the object is broken, then disabled if the object is repaired.