Addition of Mechanics to SideScroll
Created BP_Check_Side, BP_KillBox_Side & BP_PlayerSpawning_Side. They function almost the same as the split screen versions but for the side scroll characters.

Characters spawn from location of the "PlayerStart" now instead of a chosen one (can be changed to chosen location if needed)
Checkpoint respawn location can be changed, the Y value does not affect anything as they always spawn in y = 60 and y = -60

Destruction:
Destruction of objects now works in side scroll. Just stand near it and press the ability keybind and it will work if object is highlighted.
Uses event tick to figure out if player is within the area or not for whether it should be highlighted or not
Also has the KillBox to kill player if they are within the area