Manual Re-Addition of SideScroll changes
Something broke and caused issues with the original SideScroll branch meaning it could not be merged into anything without breaking a lot of the project. This is a manual re-addition of everything needed for SideScroll.

Apr 2, 2023 - DMoxom

- Created new Characters for side scroll gameplay
- Created SideScroll gamemode
- Added functionality for camera that keeps both players on the screen
- Created new swap player function as it needs to correctly assign this new camera
- Characters are restricted to Y axis. Player 1 at 60 and Player 2 at -60

Apr 28, 2023 - DMoxom

Swapped direction of the camera and inversed the controls for side-scroll gamemode so that the characters face the screen when turning

May 6, 2023 - DMoxom

Made a new box1 blueprint with different functionality for side scroll controls
Added highlight functionality when characters walk in range
Moved BP for box1 into two folders for "Side" and "Split"
Waiting before I implement the controls to actually destroy and repair the object

May 7, 2023 - Otter2701

Furlock will now enter climbing mode when they get near a wall with the "Climb" tag. Players use the W&S keys or up and down on controller to move up and down the wall players can disengage climbing by moving away from the wall

Furlock jumps if players disengage wile moving up

May 8, 2023 - DMoxom

Created BP_Check_Side, BP_KillBox_Side & BP_PlayerSpawning_Side. They function almost the same as the split screen versions but for the side scroll characters.

Characters spawn from location of the "PlayerStart" now instead of a chosen one (can be changed to chosen location if needed)
Checkpoint respawn location can be changed, the Y value does not affect anything as they always spawn in y = 60 and y = -60

Destruction:
Destruction of objects now works in side scroll. Just stand near it and press the ability keybind and it will work if object is highlighted.
Uses event tick to figure out if player is within the area or not for whether it should be highlighted or not
Also has the KillBox to kill player if they are within the area