Collectible Finish
The code which handles collectibles are now in the collectible items themselves
The Key checks if it is "Player" overlapping (works for both)
The PotShard checks if it is "Player1" overlapping it
The Fluff checks if it is "Player2" overlapping it

When the level first loads it gets all collectibles of class [key, pot, shard] and sets variables for max and remaining.
They call blueprint interface events when collected. One calls to the "CollectibleController" which will do something if collectibles left = 0.
The other calls to the widget blueprint to update the HUD with a progress bar for how far along collection is.