final changes
widgits are now focused without pressing the mouse but does not work after returning to the main menu. Key rebinding works in the main menu but if you rebind keys in the level you have to restart the level for it to take effect. The keybindings were changed from p and i and l and m to gamepad keys.